Remove the condor requirement from conductor.
As has been discussed in the past, using condor for
instance scheduling in conductor is overkill.  We can get
away with a simpler system that is easier to understand
for now, and then look at scaling up in the future.

This patch removes condor from the conductor equation and
replaces it with a combination of inline calls and dbomatic.
Specifically, when an action needs to be performed on an
instance (start, stop, etc), that call is done directly
from the conductor to the appropriate deltacloud instance.
DBomatic is then responsible for updating the database with
information about the instance as it changes state.

To do this, DBomatic has been revamped substantially.  Every
60 seconds it wakes up and attempts to update the status of
all instances.  It uses one process per provider account for
this task; this ensures that a slow or unresponsive backend
cloud will not prevent the rest of the instances from being
updated.  If a backend provider does not respond within 30
seconds, dbomatic gives up on the transaction and tries again
during the next update cycle.

This patch as it stands has been tested by me to work with the
EC2 backend.  The original patch was done by markmc with some
contributions by Slow.

Changes between v1 and v2:
a)  RHEV-M doesn't like / in names.  Convert / to - in names before
handing it off to deltacloud.  This could get dicey, as we might have
conflicting requirements from different providers.  We could either
fallback to what we did for condor (using a UUID), or we could have
a per-provider hash that has the rules.  Still to be decided.
b)  Remove DBOMATIC_PARSER_LOG_FILE, as it is no longer needed
c)  Move the import of the rails environment until *after* the daemonizing.
Apparently the classic double-fork daemonizing method wreaks havoc with
activerecord database connections, so we don't do this connection until after
we are running in the forked process
d)  Implement our own daemon (it is simple code), and move it until after
setting up the logger.  That way we find out about failures early on.
e)  Make sure to log errors while importing the rails environment or writing
the PID file to the log.

Signed-off-by: Chris Lalancette <clalance@redhat.com>