Delegate dbomatic with starting RHEV instances
Without condor, we need a mechanism to detect when RHEV instances
have been created and start them. We elected to use dbomatic.
It will start an instance if its state is STOPPED and it has
not been previously started by checking its total time in the
RUNNING state.

This patch should go in with the Taskomatic series.