Change our use of Bundler to work with Fedora.
We may be able to streamline this so there are less files to
keep in sync at some point, but the idea is that we do regular
gem requires when loading the rails app instead of having bundler
do it.  We keep the Gemfile and Gemfile.lock to allow non-fedora
contributors/users to get all the deps they need via 'bundle install',
but then do not interact with bundler any further.

The one caveat is that someone one a non-fedora platform would need
to uncomment the source line in the Gemfile, so it will pull from
rubygems.org.  In the fedora case, we dont want that - you should have
installed everything via yum/rpm, and if you are adding a new dep, this
will make it obvious to you and your tester that a new rpms dep/package
is needed before your change can be pushed to master.