Fixed links when loading data with JSON
https://www.aeolusproject.org/redmine/issues/2382
https://bugzilla.redhat.com/show_bug.cgi?id=734503

Links are now generated in views (when jquery template is defined),
and only object id is then passed in JSON data. It's not possible to pass '${id}'
string as id to *_path helper methods because URI.escape is called on id,
so after generating path, id is substituted with ('${id}').

Not saying this is cleanest way how to do it, but it's simplest ATM and buys
us some time to figure out better solution.