Make sure to always rebuild the spec files.
As jlaska points out, trying to build a non-release RPM
followed by a release one fails with the current Makefile.
That's because the spec files are listed as make targets,
which means they don't get rebuilt if they haven't changed.

Instead, revert back the old behavior where we always
unconditionally generate the spec file based on the spec.in
file.

Signed-off-by: Chris Lalancette <clalance@redhat.com>