Improved tests performance
VCR cassettes are merged to one and this one cassette is loaded only once.
There is a hack to allow other connections as it's not possible to do
it cleanly with VCR.

This is temporary solution, all tests should be updated to use
a cassette per test/scenario, not per request - cassette switing is very
expensive operation.

You can still use old style of use_cassette by setting VCR_RECORD env variable.

Signed-off-by: Matt Wagner <matt.wagner@redhat.com>