Created TemplateXML model
We need to do some operations with template such as validation
or basic editing.

Since template is warehouse/imagefactory object, we don't have
model in conductor for this. This patch adds at least basic class
where we can put all required operations (instead of having this in controller).
A template XML is validated against relax NG schema, disadvatage is that error
messages are sometimes not very user friendly.

It's partially similar to a model so I think it makes sense to put this
into app/models intead of app/utils.

Signed-off-by: Matt Wagner <matt.wagner@redhat.com>