Allow changing deployment name on the Overview page
https://bugzilla.redhat.com/show_bug.cgi?id=751024

When a deployment is launched via catalog or the deployable details page, its
name is set as the underlying deployable's name.

Previously there wasn't a way to change the name, which meant that the user
couldn't launch a second deployment using the same way.

Now the deployable overview page has a field where  the name can be changed.