BZ#755933 Last login,Last login IP fields empty.
The implementation of rails' AbstractRequest.remote_ip returns the HTTP_CLIENT_IP
if it's populated, which when there is a a proxy (apache) in front of rails, the
client ip is 127.0.0.1.  I would think rails should check the HTTP_X_FORWARDED_FOR
first, then fall back to the HTTP_CLIENT_IP.  For now, I've added a check for the
HTTP_X_FORWARDED_FOR first, then fall back to the remote_ip method call.