Fix the uptime format
https://bugzilla.redhat.com/show_bug.cgi?id=767100

The uptime output was too long when the value exceeded 1 day. Moreover, it was
not using i18n which made it untranslatable.

Now it uses standard I18n values for displaying time difference, which
incidentally makes the output long enough to fit into the pretty view
deployment boxes.

The boxes are a bit wider, too. That will hopefully make the translators' jobs
a bit easier.