Make tests more I18n-friendly
Our tests still heavily depend on the strings that we use in the English
locale. As a result, we can't run the tests on a different locale and expect
them to pass.

The changes here move us a bit closer to that goal and allow us to modify the
names of the top-level objects in the UI (like Cloud Forms is doing) without
having the tests fail.