BZ 771519 (part 3) - Added :url atribute to provider_account forms
https://bugzilla.redhat.com/show_bug.cgi?id=771519

If a provider account save fails, but this unsaved account has id set,
then 'form_for' w/o specifying explicitly url generates form's url for
update action instead of create action (it checks if id is nil or not).

Unsuccessful save with not-nil id happens if an error occurs in
after_create callback.