BZ 772996: relabel "images ready", disable 'launch' before build
https://bugzilla.redhat.com/show_bug.cgi?id=772996

If an assembly in a deployable passes XML validation, it means that
the image is a valid IWHD image, and the HWP is a valid HWP for this
conductor instance. However, the 'images ready' text is confusing
in that it implies that the images are ready to deploy, which isn't
necessarily the case. Whether the images are ready to deploy (i.e.
they're pushed) is indicated in the 'images' section.

So this commit changes 'images ready' to 'images valid' to reflect
the nature of the validation.

In addition, the 'launch' action is disabled unless all images
are pushed to at least one provider account.