Fix our JavaScript code to load under IE
IE7's JS parser has a few syntax restrictions that other browsers don't have.
This caused our JS to be invalid under IE.

Trailing coma in the object definition:

    {
      one: 1,
      two: 2,
      three: 3,
    }

This works in Chrome and Firefox (and presumably elsewhere though I didn't test
that) but not in IE7.

There we have to drop the last coma:

    {
      one: 1,
      two: 2,
      three: 3
    }

The other issue we've had is using an unquoted `class` key in objects:

Reasonable browsers:

    { class: 'a jolly old string', five: 6 }

Under IE7, this is broken and we must be using this instead:

    { 'class': 'a jolly old string', five: 6 }

I'm sure there are additional surprises waiting to trip us in the future.