BZ 750504 Editing/Deleting backend Hardware Profiles
https://bugzilla.redhat.com/show_bug.cgi?id=750504

The UI used to allow the users to edit or delete backend HWPs.

This hides the Edit and Delete buttons from the UI for the Provider Hardware
Profiles and makes sure that they can't be deleted by issuing rogue HTTP
request, etc.

Additionally, we no longer show matching profiles for backend HWPs as that
doesn't make much sense.

We may want to show all the *frontend* profiles that match a given backend one,
but that's not part of this patch.