IE: Fix the missing images in the Administer navigation
IE7 doesn't support the `:before` pseudo CSS selectors. We were using these to
provide the images.

This switches to setting the background images directly to the list item and
anchor elements. Because we can't limit the size of the images there, the
sprite file had to be modified as well.