BZ 743677 - stop/terminate instances when disabling inaccessible provider
If a provider is not accessible and has running instances, allow a user to stop them
on conductor side (no matter what is their real state).

If a provider is accessible but stop request fails for any reason, the provider is not
disabled (before the provider was disabled in any case).