BZ#781604 Fix inconsistent Search behaviour on Enter keypress
https://bugzilla.redhat.com/show_bug.cgi?id=781604

If a form has more than one submit button, it's left to the implementation
which one is going to be pressed on Enter.

It should come to no surprise to anyone that IE's behaviour is different from
all the major browsers.

This adds a little JavaScript piece that makes sure the proper button is
clicked when the user presses Enter in the search fields.