BZ#743377 Show deployable status in the pretty view
https://bugzilla.redhat.com/show_bug.cgi?id=743377

The deployment cards on the Monitor page now display a graphical status (with a
textual fallback for text-only browsers) of the deployment.

This makes it easy for the user to determine the status of the deployments at a
glance. For more detailed information (e.g. the statuses of individual
instances), the user still needs to drill down into the deployment in question.