BZ 786844 - Reporting of deleted instances
If an instance goes missing on the backend provider, we no longer ignore it.
Instead, if it's missing in two consecutive checks, we should remove it.

This is an interim solution; there is a discussion about a long-term fix here:
http://lists.fedorahosted.org/pipermail/aeolus-devel/2012-February/008735.html

Resolves https://bugzilla.redhat.com/show_bug.cgi?id=786844
(cherry picked from commit edeb5b7d94f77eb4eec911f0935f441715994dfa)