Add support for overriding default i18n file(s)
We need to be able to allow users to customize their install of
conductor to have their own names for things within the app.  This
simple patch adds that ability.  Note that it also allows for nesting
our i18n files, so we dont have to keep them in one giant file as
we do now.