BZ #795794 - Don't delete :vanished instances
If an instance goes into state 'vanished', we should leave it there,
not delete it. The deletion was meant as an interim solution until
we could handle things more gracefully, but it ended up causing more
problems than it avoided.

Related to https://bugzilla.redhat.com/show_bug.cgi?id=795794