BZ 788558 - allow terminate inaccessible instances on 'stop' action
Theis patch allows locally stopping of instances or deployments if
the instance's/deployment's provider is not accessible.

It fixes a bug in passing user param to queue_action method, also simplifies
used error messages - only (stop|reboot)_invalid_action can be raised.

This patch doesn't cover deletion of deployments which are not accessible - this is difficult
because there is a before_destroy hook on deployment which tries to destroy instances
on provider side.