BZs #798501 and #798505: Object names should be pulled from mapping file
https://bugzilla.redhat.com/show_bug.cgi?id=798505
https://bugzilla.redhat.com/show_bug.cgi?id=798501

This boils down to the following: The strings in question were NOT
being internationalized, they were pulled from the class names.  This
patch adds a mapping i18n file, which fixes both bugs (and any similar)
and future ones saying we don't internationalize the class names.

It is possible there are other cases beside permissions where this
maybe be the case, but if so, it is a simple matter of repeating the
pattern here and add a few translation strings to this file.