BZ 799314 - Clarify error when a deployable has images in the wrong environment.
It was unclear from the previous message what a user was to do. This attempts
to clarify what has gone wrong, and tries to walk the line of giving
a meaningful error that the deployable is unusable to end-users, while
giving enough detail to admins that they understand what is misconfigured.

Partially resolves https://bugzilla.redhat.com/show_bug.cgi?id=799314