BZ 800837 - Vanished instances are now included in failed?
The failed? method on an instance was not considering STATE_VANISHED
to be a failure, causing (among other things) the 'error' icon to not
display in table view. This fixes this, and has the incidental benefit
of making the code slightly more compact by using the existing FAILED_STATES
constant instead of a series of checks.

Resolves https://bugzilla.redhat.com/show_bug.cgi?id=800837