bug 790735: Enforce image permissions on deleting imported images https://bugzilla.redhat.com/show_bug.cgi?id=790735
Recent commits added most of the needed permissions checking for
various image actions. Unfortunately there was a bug in the code
that determined whether the user had permissions.

In particular, although 'and' and '&&' have a similar meaning in isolation
(i.e. 'a and b' and 'a && b' will resolve to the same result, they
have a different operator precedence. && takes precedence over assignment =
However, assignment = takes precedence over 'and'

In this case, it meant that code written like

  @foo = @bar and @bar.my_method

Did not do what was needed. The above method is equivalent to:

 (@foo = bar) and @bar.my_method

Which is obviously incorrect. It must be rewritten as:
 @foo = (@bar and @bar.my_method)
or
 @foo = @bar && @bar.my_method

In addition, not only were the 'hide the actions for non-privileged user' filters
in the view not working as desired, but the actual delete actions on provider_images_controller
and target_images_controller were not enforcing permissions.

This patch resolves both issues