BZ 796528 - instance's config params are now handled from taskomatic
Reason for moving this is that taskomatic's create_instance method automatically
sets create_failed instance's method when an exception occurs and then reraises
this exception which is handled in deployment's launch method. add_instance_config!
method raises an exceptions too and in that case we want to set create_failed state
too.