Fixes failing test
This test is failing on 1.0-product bceause it's looking for the
exact text, "Add mapping to provider", but we use the term "cloud
resource provider" there.

Since the next step attempts to click that button and references it
by its DOM ID, we can just skip this "And I should see..." step that
checks for the button, since it will fail anyway if the button is
missing.