BZ 802847 - fix hwp matching when there are multiple values for architecture
https://bugzilla.redhat.com/show_bug.cgi?id=802847

When there are multiple architectures defined in a provider hardware profile. The matching logic only uses the first listed value. This means it would only see i386 and not x86_64 when architecture is specified as ['i386', 'x86_64'].

We will begin to see multiple architectures defined for ec2 once this patch in deltacloud is released: https://git-wip-us.apache.org/repos/asf?p=deltacloud.git;a=commitdiff;h=f5b4c017f6d681594d6c7c79c039278825b67d4c

Fixed the matching logic to consider all options that may be listed in architecture.

Once both conductor and deltacloud changes are in. m1.small will be usable and matched when memory is specified as 512m and architecture is specified as x86_64.