Redmine #3211: Switch step defs for login to use DMA
https://www.aeolusproject.org/redmine/issues/3211

DMA means 'Direct Model Access'.  The basic idea is that your test
should only be testing the part of the app that the test is directly
intended for.  A good example (and the subject of this patch/feature)
is login.  The actual steps of logging in shoudl be tested in one place
only.  All other cases where the user is expected to be logged in should
avoid duplicating this set of steps, and instead assume that we have a
valid, logged in user to work with.

This concept is decribed in more detail on the wiki[1], as well as providing
a place for questions and common problems to be explained as they come up.

[1] https://www.aeolusproject.org/redmine/projects/aeolus/wiki/Testing_DMA