Redmine #3209 - fix test for new version of deltacloud client
https://www.aeolusproject.org/redmine/issues/3209

Simple fix, there will be another feature/task to add support for
verifying the provider (rather than just the deltacloud core instance),
is valid, which is now supported by the client.  In the meantime, this
removes one more case where we were hitting core directly, which we try
not to do in the tests.