Redmine #3210: Conductor should validate the provider, not just deltacloud-core instance.
https://www.aeolusproject.org/redmine/issues/3210

This patch adds a new method to the provider validation to verify
that the named provider is valid for a given deltacloud instance. In
addition, it moves this check out of the controller and down to the
model where it belongs on create action. It also adds errors to the
activerecord/en.yml, running tests on that value rather than a string.