Redmine #3319: Set up conductor to use new Bundler extension point.
https://www.aeolusproject.org/redmine/issues/3319

This is not currently 100% for upstream (ruby-only) dev, but gets us much closer.
We should now be updating only Gemfile.in, no more duplication in application.rb.

For the time being, you should be able to install everything via rpm, have the app
load from Gemfile.in (and not generate a lock file).  You should be able to start the
app, run specs and cukes, all using Gemfile.in (with no extra steps needed).  Note that
if you keep a copy of Gemfile in your src/, you will experience issues, as rails, rspec,
and cucumber look for that file to determine if they use bundler (even if we say we want to
load the deps ourselves).

The idea is that you, as a ruby developer, simply copy over the Gemfile.in to Gemfile,
and run 'bundle install'.  We still have a 'USE_BUNDLER' env var which is set to nil,
meaning for upstream you have to explicitly set it to 'yes'.  Once upstream is working
properly, we can default this to that setting, and set it to no/nil in our initscript.

If you add a dependency, once you have the gem/rpm installed, simply put it in Gemfile.in,
and bundle install using that file with:

bundle install [--gemfile=GEMFILE]