Temporary Bugfix for adding providers.
This needs to be addressed better later on, but for some reason, the
previously commited code is not working as expected for checking if a
provider has the correct name entered in conductor.  For Ec2 example,
this would verify that, say, 'us-east-1' is valid.  Currently, this
returns a 500 saying 'aws requires authorization for this request'.
Yet when I run the same command directly against the dc client in a
rails console, it works as expected.  This is just a validation, so
I propose just disabling it for now, in the interest of being able to
have an end-of-sprint minor release.  This can be revisited in the
near future.