RM #3269: Figure out what is wrong with deps (ruby/bundler mode)
Bundle install has not been working, largely because nobody has really
used it in months.  The main purpose of this patch is to get that back
to working, and the app to start.  In the process, it has added partial
fixes to get everything working on Rails 3.2.  There are some remaining
issues to be vetted by the community, but this patch gets us closer to
working there (where 'there' == non-rpm environment).

The reason for this disparity is that fedora and rhel have altered the
gemspec for 1 or more of our gem deps, so that versions that cannot coexist
in the ruby world can do so in rpmland.  The earliest version of Rails that
works with the newer rack that the rack-restful_submit gem needs is the
3.2 series.  All tests and functionality should continue to work on our
current development platforms (rpm distros), but this also gets us a good
chunk of the way to supporting the newest Rails, which we want to move to
at some point anyway.