BZ806001 - aeolus-configure will always create an admin user, need to key of a uuid not name
This change to conductor is required to allow deletion of an
existing aeolus user account. We need a rake task for it in the dc:
namespace so that we can use ActiveRecord for the deletion.

This is required so that we can easily create and delete a temporary
user account with administrative privileges during aeolus-configure runs.