bug 808031: allow Global Provider user to view Provider Accounts
https://bugzilla.redhat.com/show_bug.cgi?id=808031

For provider accounts, the 'edit' action doubles as the 'show' page.
This meant we were requiring modify privileges on the provider to even
see the provider pages. This patch relaxes the checking to reqiure
only view permissions, and in the views we hide the edit/modify actions
from users that don't have the higher privileges, and the form elements
are made read-only.

Conflicts:

	src/app/views/providers/_edit.html.haml
	src/app/views/providers/_form.html.haml
	src/app/views/providers/new.html.haml