Bug 807745 add 'use provider accounts' to global image admin role
https://bugzilla.redhat.com/show_bug.cgi?id= 807745

Image admin role has been augmented to include view and use permissions
on provider accounts. This allows global image admins to import
images from any provider account.

To see the changes you need to run 'rake dc:upgrade' which performs the
necessary data upgrade to get the role modification. The upgrade is
currently implemented inline in the rake task -- the intent is
for this task to eventually run a set of upgrade scripts from a
dedicated 'upgrades' directory, but the infrastructure for  that is not
yet set up.