Refactor SessionEntity model to work properly with API
The direct reference to the Session class in SessionEntity was causing
problems for API calls, since the session object wasn't yet in the db
(but the in-memory session_id and session data was there).

On further investigation, it's considered bad design to reference the
session class directly anyway, as it exposes too many impl details
of the session to the other models, even if it did work here.

I've refactored it to reference the session_id string in the SessionEntity
model instead, which appears to work now for both API and Web UI.