v2 RM 3317 - SysV init for conductor-delayed_job service
We support systems that are not systemd aware. Resurrected the old
SysV init script to support those systems.

Modified conductor spec file to install either the systemd or SysV
init configuration files depending on what the os supports.

v2 - updated spec file to use os version numbers instead of checking
the existence of a directory to determine if systemd is available.