Set aeolus user home directory to /usr/share/aeolus-conductor
In commit b4d148e, I modified useradd to try to create /var/aeolus as
the home directory of the aeolus user.  However, with SELinux in
enforcing mode, useradd is not allowed by policy to create directories
under /var.

This patch sets the home directory for the aeolus user to
/usr/share/aeolus-conductor and removes the -m flag from useradd,
meaning that it will not try creating that directory.  Since
/usr/share/aeolus-conductor is provided by the contents of the
aeolus-conductor rpm, in the end the aeolus user will have a valid
home directory, which is all we really need.

Signed-off-by: Steve Linabery <slinaber@redhat.com>