BZ#806001 - Create default admin via database seeds
Previously, the default admin user was created by aeolus-configure,
but this approach was confusing for users.  By creating the admin as
part of the database seeds, we are guaranteed to always have a usable
admin user at first login.  The expecation is for administrative users
to be managed via the web UI (or API), not via configure.

Signed-off-by: Steve Linabery <slinaber@redhat.com>