BZ835131 configure errors after upgrade
Add dc:admin_exists task

We need to query via ActiveRecord to see if a user with
admin privileges already exists. This allows us to replace
the behaviour in aeolus-configure that creates (and checks for)
a file to determine whether an admin has been created. Instead
we will look for an existing type of privilege.