Fix aeolus user home directory on upgrade
Previously the aeolus user's home directory was set to /var/aeolus.
This was changed to /usr/share/aeolus-conductor in commit 923f894.
However this only catches new installations; if the aeolus user
previously existed (e.g. upgrade) then the home directory would not be
updated.  This catches that case and ensures that old installations
using the old home directory are updated.

Signed-off-by: Steve Linabery <slinaber@redhat.com>