Fix broken permission migration.
The migration which adds denormalized permissions has to iterate over existing
permission records to create the DerivedPermission objects. Unfortunately,
now that the user groups code is in place, the rails permission model is
incompatible with what's in the database when the create_derived_permissions
migration is run. Normally, this sort of thing is resolved by overriding
the model classes in the migration to remove validations and other model
components that may be broken by future code changes.

In this case, we have more extensive needs at the model level, so sorting
through all of this is much more difficult. After chasing down phantom
migration failures over several hours, I decided to do the simpler thing here.
I've just moved the code that iterates over Permissions and creates Derived Permissions
from the CreateDerivedPermissions migration to the final migration that
modifies these models for user groups. Now, if a user tries to run a set of
migrations on an existing database that includes both of these, the derived
permissions are simply created later. If a user has already successfully
run the CreateDerivedPermissions migration but not the newer one, then the
 iteration over permissions to update derived permissions will be run
again, but this operation is safe to run more than once, as it compares
derived permissions in the database with what _should_ be in the database
for a given set of permissions -- so if you run it more than once, it just won't
make any db changes on subsequent runs.

to update derived permissions will