Changes to spec.in
Stylesheet compilation is done in src and the compiled files are then copied. This ensures proper asset urls in compiled stylesheets.
The converge-ui symlinks are not created in %post which caused them to remain in app_root after uninstall, they are copied from src as any other files instead.
app_root content is removed in %pre to make sure that it is clean before instalation.