fix regression on global role grants page
The recent security changes missed an 'if' condition that was present
in the global role grants page, bypassing the instantiation of the
permission object based on the classname and ID passed in, when there
was no classname or ID passed in (using global permissions object
in that case).

This commit also adds an rspec test that verifies the fix.

Signed-off-by: Jason Guiditta <jguiditt@redhat.com>