RM #3455: ruby default setup for conductor.
https://www.aeolusproject.org/redmine/issues/3455

This patch changes the default on checkout of conductor to be
consistent with expectations of upstream ruby community.  This means
we once more have a Gemfile instead of Gemfile.in.  Developers
wishing to use system rpms can simply:
  mv Gemfile Gemfile.in.

Those wishing to switch back and forth can do the above for 'system
dev', and to use system ruby, but separate gemset:
  bundle install --path bundle

Of course, you can also use rvm/gemsets or rbenv as desired to test on
a wider array of ruby vms, or to test against various dependency
versions.

Signed-off-by: Richard Su <rwsu@redhat.com>