Exclude delayed_job.log from the aeolus-conductor rpm
This file is already included in the daemons subpackage.  Under
rawhide (F18), these packages fail to install because rpm detects this
as a conflict.  I am unable to reproduce this behavior under F17, so I
suspect something has changed in the version of rpm in rawhide.

In any case, the log shouldn't be owned by both packages.