Adds an additional check in instance_errors to avoid obscure bug
In certain cases, we somehow manage to end up with an instance variable
that responds to .errors, but doesn't return an array as expected, so
calling .errors.size on it raises an exception. This patch ensures that
the errors respond to a .size method before we call it, since raising an
exception from within code to handle exceptions is not a good thing.
This patch does not address the root cause of this problem.