Improved starting of RHEVM instances (rev. 2)
RHEVM instances goes to stopped state after creating and they need
to be started explicitly. This patch improves check if explicit
start request should be sent.

It also extends stopping of RHEVM instances - it's possible that
RHEVM isntances freeze in shutting_down state on provider side
if development tools are not installed inside VM. In such case,
another 'stop' request is sent to the instance.