Point Gemfile at git repo to get image building working upstream.
This is a (likely) temporary change to point at master of aeolus-image
on github.  It is likely we can set it back to a version once we do
another bugfix release of aeolus-image to rubygems.org, should we
chosse to flip it back.  This patch should have no impact on setups
running from system ruby+rpms, just upstream+bundle dependencies.