BZ 842313 - fixed setting of deployment state when instance launch fails
If partial launch is enabled and all instances fails, the deployment
is marked as failed instead of running. Also if there is a failed instance
but rest of instances are running, the deployment is marked as incomplete.
(cherry picked from commit 5e7ae521e39ead9e1796b8b2a940896dd42087a8)