BZ 812740 - check associated images before destroying account or pool family (rev. 2)
Environment or provider account can't be deleted if there is any existing
image (or provider image) associated with them.

This required a little bit more refactoring. All destroy 'validations'
are done in before_destroy callbacks. If some callback fails, an exception
is raised so destroy is aborted and reason is in the exception message.
To avoid having rescue blocks in all controllers, new 'safe_destroy' ActiveRecord
method is defined, this method adds exception into object's errors and returns
true/false.