RM3771: Fix InstanceObserver spec for upstream
We need to test for deletion of a record, but ActiveRecord 3.2 does not
like when #reload is called on a deleted record.

With the spec in question, we can just #reload the parent record and
assert for the same thing.
(cherry picked from commit 4928032f0a799909be1e8e3b6e6e7527bf39e5f2)