RM3770: Fix Instance spec for upstream
There's no need to use FactoryGirl's `create` method in this spec, it
will test the code in question even if `build` is used.

Why the spec used to fail on upstream stack:

The `cloud_account` can be only built, not created, because the generic
`:provider_account` factory is just a stub factory meant for extending.
It doesn't assign some required attributes of ProviderAccount (e.g.
`username`), so the record cannot be saved. The F16 stack doesn't
consider this a problem when calling `create(:instance,
:provider_account => cloud_account)`, but upstream stack does.