RM3764 fix (move before_destroy callbacks)
This fixes RM3764 but also other shortcomings that didn't get caught by
specs previously.

The problem is that when before_destroy callbacks are put *below*
association definitions, they are executed *after* `destroy` is called
on associated objects that are `:dependent => :destroy`. This is
currently an open issue in Rails [1]. It causes problems with our
`destroyable?` callbacks, as their whole purpose is to prevent `destroy`
from happening if needed. They should prevent destroy from happening at
the very beginning, not halfway through.

When `destroy` only manipulates data in relational DB, the issue is not
critical, because the transaction gets rolled back. However, the objects
in memory stay in `destroyed?` state, which is what caused RM3764 spec
to fail. The objects get into correct state again when `reload` is
called on them.

The situation is more problematic when `destroy` is not fully
transactional. This is the case of Instance and InstanceKey. When
attempting to destroy undestroyable instance previously, the
`destroy_instance_key` filter on InstanceKey, which deletes an instance
key via Deltacloud API, would get called before `destroyable?` filter on
Instance. The `destroyable?` filter would roll back the transaction in
Conductor's DB, but wouldn't revert the operation with Deltacloud. This
is now fixed, as `destroyable?` gets called first.

[1] https://github.com/rails/rails/issues/3458
(cherry picked from commit cfb29e2d29e99c9863cdc61b0b649b0fd3acc844)