RM3765, RM3769: hard delete with acts_as_paranoid
It seems acts_as_paranoid hard delete feature is not compatible
with Rails 3.2. To be able to hard delete an object (`destroy!`) that
was previously soft deleted (`destroy`), the object first has to be
"undeleted" by setting `object.deleted_at = nil`.

We don't use the hard delete feature at all in Conductor, so I think we
shouldn't test for it (this commit removes parts of specs that do).

If we want to perform hard deletes in the future, we shoud rather switch
from acts_as_paranoid to something else, because the last version of
acts_as_paranoid was released 6 years ago. No wonder it doesn't work
fully with Rails 3.2 :)