additional fix for 798516: conditional logic/ruby version behavior
https://bugzilla.redhat.com/show_bug.cgi?id=798516

Apparently on f16/ruby 1.8, the conditional block with 'or' and 'not' is
not working as expected. I don't know if it's a ruby bug or something else
but it doesn't matter in this case -- using '!' and '||' is the right
way to do the conditional anyway.
(cherry picked from commit 2f1db53842a91b42b82edc65b8fd3af621f94104)